server/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database and Vector Store
chromadb>=0.4.22  # Versão compatível com Python 3.14+
sqlite-utils==3.35.1
sqlalchemy==2.0.23

# AI and ML
google-generativeai==0.3.2
# Atualizado para compatibilidade com huggingface_hub mais recente
sentence-transformers>=3.0.1,<3.2
openai==1.3.5  # Para compatibilidade com outras APIs
numpy<2.0  # Fixado para compatibilidade com chromadb e evitar upgrades quebrando ambiente

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Security and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
bcrypt==4.1.1
passlib==1.7.4
python-jose[cryptography]==3.3.0

# HTTP and File handling
httpx==0.25.2
aiofiles==23.2.1
pillow>=10.2.0  # Compatível com Python 3.14+

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0